feat: Refactor batch initialization and configuration
